test: fix relative imports in relocated admin tests
Tests moved in c5bbd89 went one directory deeper (src/<dir>/ → src/tests/<dir>/)
but the import paths kept their original two-segment relative form.
lint-staged's stash/restore cycle dropped these edits from the previous commit,
so c5bbd89 by itself doesn't compile.
